refactor: shift controller to frontend module
