+++ src/loaders/express/index.ts
@@ -14,6 +14,7 @@ import { BillingRouter } from '../../app/modules/billing/billing.routes'
 import { BounceRouter } from '../../app/modules/bounce/bounce.routes'
 import { ExamplesRouter } from '../../app/modules/examples/examples.routes'
 import { AdminFormsRouter } from '../../app/modules/form/admin-form/admin-form.routes'
+import { FrontendRouter } from '../../app/modules/frontend/frontend.server.routes'
 import { HomeRouter } from '../../app/modules/home/home.routes'
 import { MYINFO_ROUTER_PREFIX } from '../../app/modules/myinfo/myinfo.constants'
 import { MyInfoRouter } from '../../app/modules/myinfo/myinfo.routes'
@@ -27,7 +28,6 @@ import UserRouter from '../../app/modules/user/user.routes'
 import { VfnRouter } from '../../app/modules/verification/verification.routes'
 import apiRoutes from '../../app/routes'
 import { ApiRouter } from '../../app/routes/api'
-import { FrontendRouter } from '../../app/routes/frontend.server.routes'
 import * as IntranetMiddleware from '../../app/services/intranet/intranet.middleware'
 import config from '../../config/config'
